requirements.txt
questionary
colorama
simple_salesforce
PyYAML
tinydb